+++ grug/db/disk/Cargo.toml
@@ -9,11 +9,14 @@ repository    = { workspace = true }
 rust-version  = { workspace = true }
 version       = { workspace = true }
 
+[features]
+ibc = ["grug-app/ibc", "grug-jmt/ibc", "ics23"]
+
 [dependencies]
 grug-app   = { workspace = true }
-grug-jmt   = { workspace = true, features = ["ics23"] }
+grug-jmt   = { workspace = true }
 grug-types = { workspace = true }
-ics23      = { workspace = true }
+ics23      = { workspace = true, optional = true }
 rocksdb    = { workspace = true }
 tempfile   = { workspace = true }
 thiserror  = { workspace = true }
