+++ grug/db/disk/src/lib.rs
@@ -1,5 +1,7 @@
 mod db;
 mod error;
+#[cfg(feature = "ibc")]
+mod ics23;
 mod testing;
 mod timestamp;
