+++ grug/db/memory/Cargo.toml
@@ -13,7 +13,6 @@ version       = { workspace = true }
 grug-app   = { workspace = true }
 grug-jmt   = { workspace = true }
 grug-types = { workspace = true }
-ics23      = { workspace = true }
 thiserror  = { workspace = true }
 
 [dev-dependencies]
