+++ grug/jellyfish-merkle/src/lib.rs
@@ -1,10 +1,10 @@
 mod bitarray;
-#[cfg(feature = "ics23")]
+#[cfg(feature = "ibc")]
 mod ics23;
 mod node;
 mod proof;
 mod tree;
 
-#[cfg(feature = "ics23")]
+#[cfg(feature = "ibc")]
 pub use crate::ics23::*;
 pub use crate::{bitarray::*, node::*, proof::*, tree::*};
